Core builds with float

+++ Core/include/Acts/Definitions/AlgebraDefinitions.hpp
@@ -23,9 +23,8 @@
 namespace Acts {
 
 /// This is the Scalar used for all EDM and Geometry based object
-using AlgebraScalar = double;
+using AlgebraScalar = float;
 
-// Eigen definitions
 template <typename T, unsigned int rows, unsigned int cols>
 using ActsMatrix = Eigen::Matrix<T, rows, cols>;
